requirements.txt
delta-spark==3.2.0
importlib_metadata==8.0.0
py4j==0.10.9.7
pyspark==3.5.1
zipp==3.19.2
